Move AnalysisTest into a submodule to match Analysis class.
